+++ python/tests/test_variable_comparison.py
@@ -22,29 +22,29 @@ def test_comparison_op_exports_for_variable():
     _check_comparison_ops_on(var['x', :])
 
 
-def test_is_close():
+def test_isclose():
     unit = sc.units.one
     a = sc.Variable(['x'], values=np.array([1, 2, 3]), unit=unit)
-    assert sc.all(sc.is_close(a, a, 0 * unit, 0 * unit)).value
+    assert sc.all(sc.isclose(a, a, 0 * unit, 0 * unit)).value
 
 
-def test_is_close_rtol_as_number():
+def test_isclose_rtol_as_number():
     unit = sc.units.one
     a = sc.Variable(['x'], values=np.array([1, 2, 3]), unit=unit)
     # rtol as number (checks internal conversion to scipp scalar)
-    assert sc.all(sc.is_close(a, a, 0, 0 * unit)).value
+    assert sc.all(sc.isclose(a, a, 0, 0 * unit)).value
 
 
-def test_is_close_atol_defaults():
+def test_isclose_atol_defaults():
     unit = sc.units.one
     a = sc.Variable(['x'], values=np.array([1, 2, 3]), unit=unit)
-    assert sc.all(sc.is_close(a, a, rtol=0)).value
+    assert sc.all(sc.isclose(a, a, rtol=0)).value
 
 
-def test_is_close_rtol_defaults():
+def test_isclose_rtol_defaults():
     unit = sc.units.one
     a = sc.Variable(['x'], values=np.array([1, 2, 3]), unit=unit)
-    assert sc.all(sc.is_close(a, a, atol=0)).value
+    assert sc.all(sc.isclose(a, a, atol=0)).value
 
 
 def test_is_equal():