+++ variable/comparison.cpp
@@ -13,25 +13,25 @@ using namespace scipp::core;
 
 namespace scipp::variable {
 
-Variable is_close(const VariableConstView &a, const VariableConstView &b,
-                  const VariableConstView rtol, const VariableConstView atol,
-                  const NanComparisons equal_nans) {
+Variable isclose(const VariableConstView &a, const VariableConstView &b,
+                 const VariableConstView rtol, const VariableConstView atol,
+                 const NanComparisons equal_nans) {
   const auto tol = atol + rtol * abs(b);
   if (a.hasVariances() && b.hasVariances()) {
     const auto error_tol = atol + rtol * abs(variances(b));
     if (equal_nans == NanComparisons::Equal)
-      return variable::transform(a, b, tol, element::is_close_equal_nan) &
+      return variable::transform(a, b, tol, element::isclose_equal_nan) &
              variable::transform(sqrt(variances(a)), sqrt(variances(b)),
-                                 error_tol, element::is_close_equal_nan);
+                                 error_tol, element::isclose_equal_nan);
     else
-      return variable::transform(a, b, tol, element::is_close) &
+      return variable::transform(a, b, tol, element::isclose) &
              variable::transform(sqrt(variances(a)), sqrt(variances(b)),
-                                 error_tol, element::is_close);
+                                 error_tol, element::isclose);
   } else {
     if (equal_nans == NanComparisons::Equal)
-      return variable::transform(a, b, tol, element::is_close_equal_nan);
+      return variable::transform(a, b, tol, element::isclose_equal_nan);
     else
-      return variable::transform(a, b, tol, element::is_close);
+      return variable::transform(a, b, tol, element::isclose);
   }
 }